Add infrastructure to fix the negative board position problem
Currently, negative word positions break BoardLayout because it uses
arrays which can't have negative indices. This adds a 'topLeft'
member to BoardLayout representing the position of the top-left
corner, and adjusts coordinates passed in from outside so they're
now working from zero.

To reduce error, I've added a subclass of BoardPosition,
PositionAdjustedBoardPosition (TODO better name), and only that can
be used to access actual positions inside the board. It can be
constructed from an existing BoardPosition and a BoardPosition
representing the topLeft tile.

This is a mostly-pure refactoring for now.