Board generation /almost/ works; fix some bugs along the way
It can now generate boards like the one below. The problem is that
the logic to avoid adjacent intersections hasn't been enough to prevent
words being placed directly next to each other. For that, we'll need
to make the collision detection more intelligent: words can't exist
directly adjacent to the current one on any side and in any direction.

Sample board at this point:
.......c.
.....saas
.....adsa
...caused
...accues
..cuees..
duessue..
.ssea....
.e.......
.d.......