Fix bug with detecting adjacent words when new word is off board
The algorithm returned unconditionally when it should only have been
returning on true as an optimisation. This change fixes that and
updates a test which was actually cementing incorrect behaviour.

The algorithm seems to work now, in that it generates a valid
crossword - see below for an example. Indeed, it finally manages to
fit all the words into the generated board, something most of the
previous iterations haven't managed.

Next steps: improve the top-level generation algorithm to make it
generate and choose between more layouts, and hopefully generate
more interesting boards.

Sample board at this stage:

...u.......
...s.s.....
.use.u.....
...dues.ads
......a.c..
....caused.
..sea.c.s..
....u.e....
..case.....
..u.e......
cue........
..sad......