Stub code for drawing boards in Python
Set up JSON board output and a stub Python script which will take
it as input and output images.

As part of this, extract the process of converting a board to a
string into a separate interface, BoardToString, which is
implemented both by the new BoardToJson and the existing
BoardToHumanReadableString, formerly part of BoardPrinter.

Also rearrange the tests slightly to match.