Topological sort
Had some trouble with this in the past, so seeing if I can do it
properly now. Works fine.

One problem: this assumes a DAG, and won't detect cycles. The solution
is to 'soft-mark' nodes before considering their children and properly
mark them afterwards, throwing on reaching a node which is soft-marked,
but that didn't seem too important for this quick test.