Find k distinct ints in an array which sum to a given value
For k = 1 this is linear - could be binary search but it's such a rare
and irrelevant case.

For k = 2, we iterate in from each side - increase L when we're below
the target and decrease r when above. Because we sort the array first,
this will find the pair if one exists.

For higher k, iterate in from the left to fix the first int then recurse
for k-1 in the space to the right of that. Once k reaches 2 we switch to
the special case mentioned above.

Complexity = O(n^(k-1)) (with O(k) additional space assuming reasonable
garbage collection efficiency).