Match nuts to bolts
Each nut matches one bolt. We can only compare nuts to bolts (not nuts
to nuts or bolts to bolts), and we need to output a matching from nuts
to bolts.

Solution: we can divide the required space in half with each nut we
check. We compare the first nut with every bolt and thus find:
1. which bolt matches,
2. which bolts are smaller, and
3. which bolts are larger.

This lets us slowly build up a binary search tree where each node
contains a single bolt and each leaf contains a list of bolts not yet
split up. Each subsequent nut requires that we traverse the binary
search tree until we reach a leaf with a list of unprocessed bolts, at
which point we compare the nut with all the bolts to produce a node with
a value and two leaf children.

This could probably also have been done by just constructing a binary
search tree from the bolts and searching it each time, but that would
have been less silly :)