Bottom view of a binary tree
This computes the list of values which would be seen if looking at a
binary tree from below - that is, the lowest value at each horizontal
position is the one which is seen.

Horizontal positions are computed as follows:
If node n has position x, n.left has position x - 1 and n.right has
position x + 1.

If two nodes occupy the same position and depth, the one which is
rightmost in the structure of the tree will dominate.